Autocommit Thu Nov 21 20:44:16 NZDT 2013
I value kindness to human beings first of all, and kindness to animals.  I
don't respect the law; I have a total irreverence for anything connected
with society except that which makes the roads safer, the beer stronger,
the food cheaper, and old men and women warmer in the winter, and happier
in the summer.
		-- Brendan Behan